+++ src/runtime/modules/credentialstore/CredentialSelectView.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { sendError, sendResponse } from '../../utils/response'
 import { Text, View } from 'react-native'
-import YesNo from '../../utils/YesNo'
+import YesNo from '../../../components/YesNo'
 import type { KiltCredential } from '../../interfaces'
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
